// frontend/robo-app/src/services/types/roboDto.ts
export type roboDto = {
  cabecaRotacao?: string;
  cabecaInclinacao?: string;
  bracoDireitoCotoveloContracao?: string;
  bracoDireitoPulsoRotacao?: string;
  bracoEsquerdoCotoveloContracao?: string;
  bracoEsquerdoPulsoRotacao?: string;
  proximosCabecaRotacao?: string[];
  proximosCabecaInclinacao?: string[];
  proximosBracoDireitoCotoveloContracao?: string[];
  proximosBracoDireitoPulsoRotacao?: string[];
  proximosBracoEsquerdoCotoveloContracao?: string[];
  proximosBracoEsquerdoPulsoRotacao?: string[];
};